Fix for window function now taking long param instead of int.

+++ Editor/EditorCore/UVEditor.cs
@@ -2508,7 +2508,7 @@ namespace UnityEditor.ProBuilder
         static Rect ActionWindowDragRect = new Rect(0, 6, 10000, 30);
         static Editor uv2Editor = null;
 
-        void DrawActionWindow(int windowIndex)
+        void DrawActionWindow(long windowIndex)
         {
             if (channel == 0)
             {
